Add SOLVE function, using ZGESV
Solves system of linear equations. Chose the complex-valued option
ZGESV because that seems to be a good choice / the ZGE-prefixed LAPACK
functions are used elsewhere in the code. Not sure if the other LAPACK
solvers should be exposed?

Produces the same results as shown in:
https://software.intel.com/sites/products/documentation/doclib/mkl_sa/11/mkl_lapack_examples/zgesv_ex.c.htm

(setq a (make-complex-matrix
         4 4
         '(#C(1.23 -5.50) #C(-2.14 -1.12) #C(-4.30 -7.10) #C( 1.27  7.29)
           #C( 7.91 -5.38) #C(-9.92 -0.79) #C(-6.47  2.52) #C( 8.90  6.92)
           #C(-9.80 -4.86) #C(-9.18 -1.12) #C(-6.51 -2.67) #C(-8.82  1.25)
           #C(-7.32  7.57) #C( 1.37  0.43) #C(-5.86  7.38) #C( 5.41  5.37))))

(setq b (make-complex-matrix
         4 2
         '(#C( 8.33 -7.32) #C(-6.18 -4.80) #C(-5.71 -2.80) #C(-1.60  3.08)
           #C(-6.11 -3.81) #C( 0.14 -7.71) #C( 1.41  3.40) #C( 8.54
         -4.05))))

(solve a b)