nicer error for NO-APPLICABLE-IMPLEMENTATION
The old error was quote opaque. It would say something like

```
MAGICL:EXPM is not implemented in any of the current active backends: EXPOKIT, LAPACK, BLAS, LISP.
[Condition of type MAGICL.BACKENDS:NO-APPLICABLE-IMPLEMENTATION]
```

Now it says:

```
No applicable implementations found for the supposed backend function MAGICL:EXPM.

MAGICL:EXPM names a backend function.

Active Backends:
    * Backend :EXPOKIT has MAGICL-EXPOKIT::EXPM-EXPOKIT [generic function]
    * Backend :LAPACK has no implementations
    * Backend :BLAS has no implementations
    * Backend :LISP has no implementations

There are no inactive backends.

There was no implementation of MAGICL:EXPM that was applicable to the given arguments. This may be because:

    * No implementation exists in a currently active backend,

    * A generic function implements the backend, but the generic function does not specialize on the argument, or

    * An implementation exists, but signaled that it wasn't applicable for the given arguments.

Ensure the proper backends are activated, and the implementations in those backends can handle the argument types appropriately. If an implementation does not exist, consider writing one!
   [Condition of type MAGICL.BACKENDS:NO-APPLICABLE-IMPLEMENTATION]
```