Resolves #156: Cannot load fortran lib
File changed: src/extensions/expokit/load-libs.lisp

This changes fortran library loading to correspond with the recent
changes in magicl.asd that changed the directory location of the
LIBEXPOKIT fortran library.

Previously, the loading code defined the library to exist in a
directory relative to the source tree, and now it defines that
directory relative to the output tree, i.e., where we output fasls,
etc., as per the recent change in magicl.asd.

We now use cffi:*foreign-library-directories* to define the location
instead of sticking read-time-evaluated pathnames into the library
definition. It's less code, since it does not need to branch per
platform (cf. old darwin/linux cases). It also should work on Windows
besides macOS and Linux, though that hasn't been tested. It also is
better in that the location setting is done at load time, whereas the
old approach did it at read time (not really the right time).

Along the way, noticed and fixed a minor problem:
*expokit-libs-loaded* was not being set to true after loading.